# docs/api/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "1Cati Site Management API",
    "version": "2026.07.01",
    "summary": "API contract for the 1Cati property-management web application.",
    "description": "This OpenAPI contract describes the local and production-ready API surface for the 1Cati CRM/ERP dashboard. It is used for API testing, Jira/Xray traceability and future client/provider handover."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3104",
      "description": "Local development server"
    },
    {
      "url": "https://cati-blond.vercel.app",
      "description": "Hosted web application"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "AccessProfileCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_profile_role",
        "description": "Local QA access-profile cookie. Production uses Supabase Auth instead."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" }
        }
      },
      "SourceResponse": {
        "type": "object",
        "properties": {
          "source": { "type": "string", "enum": ["supabase", "local-seed"] }
        },
        "additionalProperties": true
      },
...
      "ActionLogRequest": {
        "type": "object",
        "required": ["actionType"],
        "properties": {
          "actionType": { "type": "string", "maxLength": 120 },
          "entityTable": { "type": "string" },
          "entityId": { "type": "string" },
          "entityExternalId": { "type": "string" },
          "title": { "type": "string" },
          "metadata": { "type": "object", "additionalProperties": true }
        }
      },
      "ActionDecisionRequest": {
        "type": "object",
        "required": ["id", "status"],
        "properties": {
          "id": { "type": "string", "minLength": 1 },
          "status": { "type": "string", "enum": ["approved", "rejected", "completed", "failed"] },
          "actionType": { "type": "string", "maxLength": 120, "default": "ticket.update.request" },
          "entityTable": { "type": "string", "default": "service_tickets" }
        },
        "additionalProperties": false
      },
      "TicketCreateRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": { "type": "string", "minLength": 3, "maxLength": 160 },
          "description": { "type": "string", "maxLength": 1200 },
          "category": { "type": "string", "default": "general" },
          "unitNo": { "type": "string" },
          "priority": { "type": "string", "enum": ["low", "normal", "high", "urgent"], "default": "normal" },
          "origin": { "type": "string", "enum": ["api", "ai"], "default": "api" }
        },
        "additionalProperties": true
      },
...
      "TicketCreateResponse": {
        "type": "object",
        "properties": {
          "workflow": { "type": "object", "additionalProperties": true },
          "ticketRequest": {
            "type": "object",
            "properties": {
              "status": { "type": "string", "enum": ["submitted"] },
              "proposedPayload": { "type": "object", "additionalProperties": true },
              "requiresHumanApproval": { "type": "boolean" }
            },
            "additionalProperties": true
          }
        },
        "additionalProperties": true
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "The user is not authenticated.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Forbidden": {
        "description": "The role is not allowed to perform this action.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "ServerError": {
        "description": "The requested data or action is currently unavailable.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  },
  "security": [{ "AccessProfileCookie": [] }],
  "paths": {
...
    "/api/site-management/tickets": {
      "get": {
        "tags": ["Operations"],
        "operationId": "getTicketsAndTasks",
        "summary": "Return tickets, service orders and workforce tasks.",
        "description": "Returns service ticket, task and SLA data for operations workflows.",
        "parameters": [{ "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100 } }],
        "responses": { "200": { "description": "Ticket data returned." }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      },
      "post": {
        "tags": ["Operations"],
        "operationId": "createServiceTicketRequest",
        "summary": "Submit a service ticket request for human-approved workflow processing.",
        "description": "Creates an audited service ticket request from API or AI origin. The backend validates role permission, title length, description length and approval workflow metadata before returning a submitted request.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TicketCreateRequest" }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Ticket request submitted for workflow processing.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TicketCreateResponse" }
              }
            }
          },
          "400": { "description": "Invalid ticket body or field length.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/api/site-management/actions": {
      "post": {
        "tags": ["Operations"],
        "operationId": "logClientAction",
        "summary": "Log an audited client action.",
        "description": "Stores an audited action after validating role permissions against the target entity and action type.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ActionLogRequest" } } } },
        "responses": {
          "201": { "description": "Action logged.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SourceResponse" } } } },
          "400": { "description": "Invalid action body." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "patch": {
        "tags": ["Operations"],
        "operationId": "decideClientActionRequest",
        "summary": "Approve or reject an audited client action request.",
        "description": "Updates a pending action request after validating that the current role has approval permission for the resolved workflow action.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ActionDecisionRequest" } } } },
        "responses": {
          "200": { "description": "Action request decision stored.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SourceResponse" } } } },
          "400": { "description": "Invalid action decision body." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  }
}